backend/requirements.txt
adrf==0.1.6
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
multidict==6.0.5
packaging==24.1
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
